Feature/add integration tests (#48)
Sorry, it's a big PR, but there was just a lot of template files moving
around. Most of the important changes are in `cli/commands` and
`core/lib`

When compiled by `cargo contract` or ask compiler, and then
type-generated by typechain-polkadot package, the TS typed contracts
don't work properly and is not possible to run tests.

This was solved by replacing cargo contract by typechain-compiler.
Result is that we can for now only have working integration tests for
ink! based contracts.

This will probbably be fixed by syncing all the polkadot.js dependency
versions with typechain side.

Notable changes in this PR:

- added tests to templates
- changed the template folder structure
- replaced `cargo contract` with `typechain-compiler`
- added renaming and move steps to compile step to be compatible with
typechan outputs

Co-authored-by: PierreOssun <pierre.giraud@mailfence.com>
Co-authored-by: PierreOssun <35110271+PierreOssun@users.noreply.github.com>